ShoopedImp broken up into separate files
- Made Type::for public
